convert painter and other render classes to ES6 class syntax
